api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
python-dotenv==1.0.1
gunicorn==22.0.0
python-dotenv
